Refactor and add functions
