+++ src/utils/image.js
@@ -31,6 +31,30 @@ export const fileToBase64 = file => {
 }
 
 /**
+ * Converts data-uri image to file/Blob
+ * Source: https://stackoverflow.com/questions/4998908/convert-data-uri-to-file-then-append-to-formdata
+ * @function base64ToBlob
+ * @param {String}
+ * @return {File}
+ */
+export const base64ToBlob = string => {
+  if (string instanceof Blob) return string
+  let byteString
+  if (string.split(",")[0].indexOf("base64") >= 0)
+    byteString = atob(string.split(",")[1])
+  else byteString = unescape(string.split(",")[1])
+  const mimeString = string
+    .split(",")[0]
+    .split(":")[1]
+    .split(";")[0]
+  let ia = new Uint8Array(byteString.length)
+  for (let i = 0; i < byteString.length; i++) {
+    ia[i] = byteString.charCodeAt(i)
+  }
+  return new Blob([ia], {type: mimeString})
+}
+
+/**
  * Enforces image size and filetype.
  * @function forceImageRestrictions
  * @param {Int} size Size in bytes.
