+++ src/utils/index.js
@@ -1,15 +1,17 @@
-export {focusEvents} from "./events-focus"
-export {addLink, formatCommand, menuPosition} from "./actions-format"
+export {focusEvents} from "./focus"
+export {addLink, formatCommand, menuPosition} from "./format"
 export {
   imageButtonPosition,
   handleImageButton,
   handleFileUpload,
   forceImageRestrictions,
-} from "./actions-image"
+  fileToBase64,
+  base64ToBlob,
+} from "./image"
 export {
   loadContent,
   loadTextContent,
   storeContentState,
   saveContent,
   setDraftStatusHelper,
-} from "./actions-storage"
+} from "./storage"
